Put include files in an include directory
