+++ include/otdgen/DocBlocks.h
@@ -13,7 +13,7 @@
 
 /*\\\ INCLUDE FILES \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*/
 
-#include "DocBlock.h"
+#include "otdgen/DocBlock.h"
 
 #include <list>
 #include <memory>
