+++ include/otdgen/DocParameter.h
@@ -1,6 +1,6 @@
 /*****************************************************************************
 
-      DocCartouche.h
+      DocParameter.h
       Copyright (c) 2015 Raphael DINGE
 
 *Tab=3***********************************************************************/
@@ -13,8 +13,10 @@
 
 /*\\\ INCLUDE FILES \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*/
 
-#include "DocBlock.h"
-#include "DocInline.h"
+#include "otdgen/DocBlock.h"
+#include "otdgen/DocInline.h"
+
+#include <list>
 
 
 
@@ -23,7 +25,7 @@ namespace otdgen
 
 
 
-class DocCartouche
+class DocParameter
 :  public DocBlock
 {
 
@@ -31,20 +33,16 @@ class DocCartouche
 
 public:
 
-                  DocCartouche () = default;
-                  DocCartouche (const DocCartouche & rhs) = default;
-                  DocCartouche (DocCartouche && rhs) = default;
-   virtual        ~DocCartouche () = default;
+                  DocParameter () = default;
+                  DocParameter (const DocParameter & rhs) = default;
+                  DocParameter (DocParameter && rhs) = default;
+   virtual        ~DocParameter () = default;
 
-   DocCartouche & operator = (const DocCartouche & rhs) = default;
-   DocCartouche & operator = (DocCartouche && rhs) = default;
+   DocParameter & operator = (const DocParameter & rhs) = default;
+   DocParameter & operator = (DocParameter && rhs) = default;
 
-   std::string    inherit;
-   std::string    header;
-   std::string    guide_id;
-   DocInlines     guide;
-   std::list <std::string>
-                  declarations;
+   std::string    type;
+   DocInlines     body;
 
 
 
@@ -67,20 +65,21 @@ private:
 /*\\\ FORBIDDEN MEMBER FUNCTIONS \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*/
 
 private:
-   bool           operator == (const DocCartouche & rhs);
-   bool           operator != (const DocCartouche & rhs);
+   bool           operator == (const DocParameter & rhs);
+   bool           operator != (const DocParameter & rhs);
 
 
 
-}; // class DocCartouche
+}; // class DocParameter
 
+using DocParameters = std::list <DocParameter>;
 
 
 }  // namespace otdgen
 
 
 
-// #include "DocCartouche.hpp"
+// #include "DocParameter.hpp"
