+++ include/otdgen/GeneratorGitHubMarkDown.h
@@ -13,11 +13,11 @@
 
 /*\\\ INCLUDE FILES \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*/
 
-#include "DocInline.h"
-#include "DocMember.h"
-#include "DocParameter.h"
-#include "DocType.h"
-#include "GeneratorBase.h"
+#include "otdgen/DocInline.h"
+#include "otdgen/DocMember.h"
+#include "otdgen/DocParameter.h"
+#include "otdgen/DocType.h"
+#include "otdgen/GeneratorBase.h"
 
 #include <string>
 #include <vector>
