+++ include/otdgen/LexicalAnalyser.h
@@ -13,7 +13,7 @@
 
 /*\\\ INCLUDE FILES \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*/
 
-#include "Token.h"
+#include "otdgen/Token.h"
 
 #include <list>
 #include <map>
