+++ include/otdgen/StructuralAnalyser.h
@@ -13,7 +13,7 @@
 
 /*\\\ INCLUDE FILES \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*/
 
-#include "DocLibrary.h"
+#include "otdgen/DocLibrary.h"
