+++ include/otdgen/SyntaxicAnalyser.h
@@ -13,8 +13,8 @@
 
 /*\\\ INCLUDE FILES \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*/
 
-#include "Expression.h"
-#include "Token.h"
+#include "otdgen/Expression.h"
+#include "otdgen/Token.h"
 
 #include <map>
 #include <string>
