+++ otdgen_src.gypi
@@ -1,0 +1,24 @@
+##############################################################################
+#
+#     otdgen_src.gypi
+#     Copyright (c) 2014 Raphael DINGE
+#
+#Tab=3########################################################################
+
+
+
+{
+   'include_dirs': [
+      'include',
+   ],
+
+   'direct_dependent_settings': {
+      'include_dirs': [
+         'include',
+      ],
+   },
+
+   'includes': [
+      'src/src.gypi',
+   ],
+}
