+++ src/Conf.cpp
@@ -9,7 +9,7 @@
 
 /*\\\ INCLUDE FILES \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*/
 
-#include "Conf.h"
+#include "otdgen/Conf.h"
 
 #include <map>
