+++ src/DocBlocks.cpp
@@ -9,7 +9,7 @@
 
 /*\\\ INCLUDE FILES \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*/
 
-#include "DocBlocks.h"
+#include "otdgen/DocBlocks.h"
 
 #include <cassert>
