+++ src/DocBook.cpp
@@ -9,7 +9,7 @@
 
 /*\\\ INCLUDE FILES \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*/
 
-#include "DocBook.h"
+#include "otdgen/DocBook.h"
 
 #include <cassert>
