+++ src/DocChapter.cpp
@@ -9,7 +9,7 @@
 
 /*\\\ INCLUDE FILES \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*/
 
-#include "DocChapter.h"
+#include "otdgen/DocChapter.h"
 
 #include <cassert>
